requirements.txt
aiohttp>=3.9,<4.0
Pillow>=10.0.0

